Move, restructured directories for lightning integration
