+++ pyproject.toml
@@ -7,7 +7,6 @@ requires-python = ">=3.12"
 dependencies = [
     "loguru>=0.7.3",
     "matplotlib>=3.10.8",
-    "pytorch-lightning>=2.6.1",
     "torch>=2.10.0",
     "tqdm>=4.67.3",
 ]
