+++ src/hand_to_tex/models/lit_module.py
@@ -1,4 +1,4 @@
-import lightning as pl
+import lightning.pytorch as pl
 import torch.nn as nn
 import torch.optim as optim
 from torch import Tensor
