--- src/hand_to_tex/training/__init__.py
@@ -1,4 +1,0 @@
-from .baseline_trainer import BaselineTrainer
-from .lightning_trainer import HMELightningModule
-
-__all__ = ["BaselineTrainer", "HMELightningModule"]
